Handling the lack of a holdings directory a little more nicely.

+++ core/management/commands/load_holdings.py
@@ -1,4 +1,5 @@
 import logging
+import os
 
 from django.core import management
 from django.core.management.base import BaseCommand
@@ -24,10 +25,18 @@ class Command(BaseCommand):
 
     def handle(self, holdings_source=default_location, *args, **options):
         
+        if not os.path.exists(holdings_source): 
+            _logger.error("There is no valid holdings source folder defined.")
+            set_holdings = ['To load holdings - Add a folder called "holdings"',
+            'to the bib directory that is set in settings',
+            'or pass the location of holdings as an arguement to the loader.',]
+            _logger.error(' '.join(set_holdings))
+            return
+
         # First we want to make sure that our material types are up to date
         material_types = models.MaterialType.objects.all()
         [m.delete() for m in material_types]
         management.call_command('loaddata', 'material_types.json')
-        
+       
         holding_loader = HoldingLoader()
         holding_loader.main(holdings_source)
